Testing redis things
Currently testing using Redis for all communication. \nRedis store for connected workers, another service for ensuring all workers in store are online, each worker listens on a channnel prefixed with a uuid for that worker, gateway selects best worker for the event and sents on the correct channel. I shouldnt need workers to communicate back to the gateway but in the future if that is needed then the same logic can be used just in reverse\n\n\nAlso does \n work in git commits from cli?